Revert solution type changes
