+++ pallets/pool-system/src/solution.rs
@@ -10,13 +10,18 @@
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU General Public License for more details.
 
-use cfg_types::{epoch::UnhealthyState, pools::PoolState};
 use frame_support::sp_runtime::traits::Convert;
 use sp_arithmetic::traits::Unsigned;
 use sp_runtime::ArithmeticError;
 use sp_std::vec;
 
 use super::*;
+
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+pub enum PoolState {
+	Healthy,
+	Unhealthy(Vec<UnhealthyState>),
+}
 
 impl PoolState {
 	/// Updates a PoolState to update.
@@ -77,6 +82,21 @@ impl PoolState {
 			}
 		}
 	}
+}
+
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub enum UnhealthyState {
+	MaxReserveViolated,
+	MinRiskBufferViolated,
+}
+
+/// The solutions struct for epoch solution
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub enum EpochSolution<Balance> {
+	Healthy(HealthySolution<Balance>),
+	Unhealthy(UnhealthySolution<Balance>),
 }
 
 impl<Balance> EpochSolution<Balance> {
@@ -281,6 +301,13 @@ where
 	}
 }
 
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub struct HealthySolution<Balance> {
+	pub solution: Vec<TrancheSolution>,
+	pub score: Balance,
+}
+
 impl<Balance> PartialOrd for HealthySolution<Balance>
 where
 	Balance: PartialOrd,
@@ -288,6 +315,17 @@ where
 	fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
 		self.score.partial_cmp(&other.score)
 	}
+}
+
+#[derive(Encode, Decode, Clone, Eq, PartialEq, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub struct UnhealthySolution<Balance> {
+	pub state: Vec<UnhealthyState>,
+	pub solution: Vec<TrancheSolution>,
+	// The risk buffer score per tranche (less junior tranche) for this solution
+	pub risk_buffer_improvement_scores: Option<Vec<Balance>>,
+	// The reserve buffer score for this solution
+	pub reserve_improvement_score: Option<Balance>,
 }
 
 impl<Balance> UnhealthySolution<Balance> {
@@ -346,6 +384,14 @@ where
 		// If both of the above rules to not apply, we value the solutions as equal
 		Some(Ordering::Equal)
 	}
+}
+
+// The solution struct for a specific tranche
+#[derive(Encode, Decode, Copy, Clone, Eq, PartialEq, Default, RuntimeDebug, TypeInfo)]
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+pub struct TrancheSolution {
+	pub invest_fulfillment: Perquintill,
+	pub redeem_fulfillment: Perquintill,
 }
 
 pub fn calculate_solution_parameters<Balance, BalanceRatio, Rate, Weight, Currency>(
